Add LittleFS support
This will be usefull for config like tags and participants.

This can be used to save a (ongoing) race and load it again.

Of possible it should be possible to powercycle/crach the device during
a race without loosing anything. It should also be possible to download
race date when race is over.

Signed-off-by: Zingo Andersen <zingo@zingo.org>