Refactor race and tag data into it's own task
BT task send a message to it when it sees a tag and it does all logic
If it sees a tag for the first time it send a message back to the BT
task to make it handle the first connect and setup, then it send a
message back when done.

This disconect the knwoladge and handling of the data structures and
make different parts of the system more isolated and easier to keep
protected. It also will make some testing down the line easier as
we can just fabricate the message that the tags would send. Or if
we add a second unit it can act as a dumb unit and just send the itag
messages to the head unit, this could be used to make the BT scanning
better by adding units.

Signed-off-by: Zingo Andersen <zingo@zingo.org>