Split gui into race and participants lists
Race tab:
Contains participants that are part of the current race
This will show race info like distance, laps and time

Participant tab:
All tags, Idea is that you can edit name and select
who is part of the race.
This will show and edit user and Tag info, maybe show stats.

This change also added info on who is part in the race and populate the
race tab via messages to start the task to separate UI and Race DB data
and info.

Signed-off-by: Zingo Andersen <zingo@zingo.org>