RaceDB Thread handle it's own timing
This change will setup a FreeRTOS 2000ms repeated timer and use it
internally in the RaceDB code, this moves the running code to handle
the UI and data structure updates each 2000ms from main thread to
RaceDB context.

This solves the problem with multiple contextes accessing the RaceDB
data and let us remove the locking mutex and might avoid deadlocks.

Signed-off-by: Zingo Andersen <zingo@zingo.org>