Add goal and show it in graph
Add More info about a participant like pace from start, now and rest to
make a goal.
Add personal goal that can be chanegd during race in graph view.
Plot a goal in graph so it's easy to check.
Improve graph view to autoscale to 20% above goal and change if goal is
changed.

Signed-off-by: Zingo Andersen <zingo@zingo.org>