Defaul to only show one user for now
As current code is sligtly aimed for my VXO SM 24H 2023 race lets keep the default
setting simple and aimed for it for now. Graph and Tags will be hardcoded for a few version now. Plan is to make it configurable but for now it follows
DEFAULT_PARTICIPANT in common.h